backend/requirements.txt
# Web框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# 数据库
sqlalchemy>=2.0.0
alembic>=1.12.0
aiosqlite>=0.19.0

# HTTP客户端
httpx>=0.25.0
aiohttp>=3.9.0

# 数据处理
pandas>=2.1.0
numpy>=1.24.0
akshare>=1.13.0
aktools>=0.0.90

# 任务调度
apscheduler>=3.10.0

# WebSocket
python-socketio>=5.9.0

# 配置管理
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# 数据验证
pydantic>=2.4.0

# 日志
loguru>=0.7.0

# 工具库
python-dateutil>=2.8.0
python-multipart>=0.0.6

# 开发工具
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0

# 测试工具
httpx>=0.25.0
pytest-cov>=4.1.0